Consolidate hooks into single directory
