+++ apps/website/src/components/shared/form/UploadAttachmentsField.tsx
@@ -8,11 +8,15 @@ import {
   type ChangeEvent,
   type Key,
 } from "react";
+
 import { fileToBase64 } from "@/utils/files";
+
 import IconUploadFiles from "@/icons/IconUploadFiles";
+
+import useFileDragAndDrop from "@/hooks/files/drop";
+
 import { Button, defaultButtonClasses } from "../Button";
 import { MessageBox } from "../MessageBox";
-import { useFileDragAndDrop } from "../hooks/useFileDragAndDrop";
 
 export type FileReference =
   | InitialFileReference
