+++ apps/website/src/components/show-and-tell/ShowAndTellNavigation.tsx
@@ -8,7 +8,7 @@ import {
 
 import { classes } from "@/utils/classes";
 
-import { useIsActivePath } from "@/components/shared/hooks/useIsActivePath";
+import useIsActivePath from "@/hooks/active";
 
 type NavLinkProps = Omit<
   AnchorHTMLAttributes<HTMLAnchorElement>,
