+++ apps/website/src/pages/show-and-tell/index.tsx
@@ -15,6 +15,9 @@ import { delay } from "@/utils/delay";
 import { trpc } from "@/utils/trpc";
 import { getPosts } from "@/server/db/show-and-tell";
 
+import useOnToggleNativeFullscreen from "@/hooks/fullscreen";
+import useIntersectionObserver from "@/hooks/intersection";
+
 import IconLoading from "@/icons/IconLoading";
 import IconArrowUp from "@/icons/IconArrowUp";
 import IconArrowDown from "@/icons/IconArrowDown";
@@ -23,8 +26,6 @@ import IconArrowsOut from "@/icons/IconArrowsOut";
 import IconPencil from "@/icons/IconPencil";
 
 import { Button, LinkButton } from "@/components/shared/Button";
-import { useOnToggleNativeFullscreen } from "@/components/shared/hooks/useOnToggleNativeFullscreen";
-import { useIntersectionObserver } from "@/components/shared/hooks/useIntersectionObserver";
 
 import Meta from "@/components/content/Meta";
 import Section from "@/components/content/Section";
